+++ webapp/apps/dynamic/tests/test_elasticity.py
@@ -4,6 +4,7 @@ from django.test import Client
 import mock
 import os
 import json
+import sys
 # os.environ["NUM_BUDGET_YEARS"] = '2'
 
 from ...taxbrain.models import TaxSaveInputs
@@ -13,7 +14,7 @@ from ...taxbrain.compute import DropqCompute, MockCompute, ElasticMockCompute
 import taxcalc
 from taxcalc import Policy
 
-from .utils import do_elasticity_sim, START_YEAR
+from .utils import do_dynamic_sim, START_YEAR
 from ...test_assets.utils import (check_posted_params, do_micro_sim,
                                   get_post_data, get_file_post_data)
 
@@ -29,7 +30,6 @@ class DynamicElasticityViewsTests(TestCase):
         self.client = Client()
 
     def test_elasticity_edit(self):
-        import sys
         from webapp.apps.taxbrain import views
         webapp_views = sys.modules['webapp.apps.taxbrain.views']
         webapp_views.dropq_compute = MockCompute()
@@ -55,26 +55,47 @@ class DynamicElasticityViewsTests(TestCase):
 
         # Do the elasticity of GDP simulation based on the third microsim
         egdp_reform = {u'elastic_gdp': [u'0.4']}
-        egdp_response = do_elasticity_sim(self.client, micro3, egdp_reform)
+        egdp_response = do_dynamic_sim(self.client, 'macro', micro3, egdp_reform)
         orig_micro_model_num = micro3.url[-2:-1]
 
         # Now edit this elasticity of gdp sim
         # Go to macro input page
         egdp_num = egdp_response.url[egdp_response.url[:-1].rfind('/')+1:-1]
-        dynamic_macro_edit = '/dynamic/macro/edit/{0}?start_year={1}'.format(egdp_num, START_YEAR)
-        #Redirect first
+        dynamic_macro_edit = '/dynamic/macro/edit/{0}/?start_year={1}'.format(egdp_num, START_YEAR)
+        # get edit page.
         response = self.client.get(dynamic_macro_edit)
-        self.assertEqual(response.status_code, 301)
-        #Now load the page
-        rep2 = self.client.get(response.url)
-        page = rep2.content
+        self.assertEqual(response.status_code, 200)
+        page = response.content
         # Read the page to find the linked microsim. It should be the third
         # microsim above
         idx = page.find('dynamic/macro')
         idx_ms_num_start = idx + 14
         idx_ms_num_end = idx_ms_num_start + page[idx_ms_num_start:].find('/')
         microsim_model_num = page[idx_ms_num_start:idx_ms_num_end]
-        assert microsim_model_num == orig_micro_model_num
+        microsim_url = page[idx:idx_ms_num_end]
+        assert 'dynamic/macro/{0}'.format(microsim_model_num) == microsim_url
+
+    def test_elasticity_cps(self):
+        # Do the microsim
+        start_year = u'2016'
+        data = get_post_data(start_year)
+        data[u'II_em'] = [u'4333']
+        data['data_source'] = ['CPS']
+
+        micro1 = do_micro_sim(self.client, data)
+
+        from webapp.apps.dynamic import views
+        dynamic_views = sys.modules['webapp.apps.dynamic.views']
+        dynamic_views.dropq_compute = ElasticMockCompute(num_times_to_wait=1)
+
+        # Do the elasticity of GDP simulation based on the third microsim
+        egdp_reform = {u'elastic_gdp': [u'0.4']}
+        egdp_response = do_dynamic_sim(self.client,'macro', micro1['response'],
+                                          egdp_reform)
+        from webapp.apps.dynamic import views
+        post = views.dropq_compute.last_posted
+
+        assert post['use_puf_not_cps'] == False
 
     @pytest.mark.xfail
     def test_elasticity_reform_from_file(self):
@@ -97,7 +118,7 @@ class DynamicElasticityViewsTests(TestCase):
 
         # Do the partial equilibrium simulation based on the microsim
         el_reform = {u'elastic_gdp': [u'0.4']}
-        el_response = do_elasticity_sim(self.client, micro1, el_reform)
+        el_response = do_dynamic_sim(self.client, 'macro', micro1, el_reform)
         orig_micro_model_num = micro1.url[-2:-1]
         from webapp.apps.dynamic import views
         post = views.dropq_compute.last_posted