+++ webapp/apps/dynamic/tests/utils.py
@@ -6,7 +6,7 @@ import os
 import sys
 # os.environ["NUM_BUDGET_YEARS"] = '2'
 
-from ...taxbrain.models import TaxSaveInputs
+from ...taxbrain.models import TaxSaveInputs, OutputUrl
 from ...taxbrain.helpers import (expand_1D, expand_2D, expand_list, package_up_vars,
                                  format_csv, arrange_totals_by_row, default_taxcalc_data)
 from ...taxbrain.compute import DropqCompute, MockCompute, ElasticMockCompute
@@ -20,15 +20,16 @@ from ...test_assets.utils import get_dropq_compute_from_module, do_micro_sim
 START_YEAR = u'2016'
 
 
-def do_behavioral_sim(client, microsim_response, pe_reform, start_year=START_YEAR):
+def do_dynamic_sim(client, base_name, microsim_response, pe_reform, start_year=START_YEAR):
     # Link to dynamic simulation
-    model_num = microsim_response.url[-2:]
-    dynamic_landing = '/dynamic/{0}?start_year={1}'.format(model_num, start_year)
+    idx = microsim_response.url[:-1].rfind('/')
+    model_num = microsim_response.url[idx+1:-1]
+    dynamic_landing = '/dynamic/{1}/?start_year={2}'.format(base_name, model_num, start_year)
     response = client.get(dynamic_landing)
     assert response.status_code == 200
 
     # Go to behavioral input page
-    dynamic_behavior = '/dynamic/behavioral/{0}?start_year={1}'.format(model_num, start_year)
+    dynamic_behavior = '/dynamic/{0}/{1}/?start_year={2}'.format(base_name, model_num, start_year)
     response = client.get(dynamic_behavior)
     assert response.status_code == 200
 
@@ -48,32 +49,7 @@ def do_behavioral_sim(client, microsim_response, pe_reform, start_year=START_YEA
             reload_count = 0
         else:
             raise RuntimeError("unable to load results page")
-
-    assert response.url[:-2].endswith("behavior_results/")
-    return response
-
-
-def do_elasticity_sim(client, microsim_response, egdp_reform, start_year=START_YEAR):
-    # Link to dynamic simulation
-    model_num = microsim_response.url[-2:]
-    dynamic_landing = '/dynamic/{0}?start_year={1}'.format(model_num, start_year)
-    response = client.get(dynamic_landing)
-    assert response.status_code == 200
-
-    # Go to macro input page
-    dynamic_egdp = '/dynamic/macro/{0}?start_year={1}'.format(model_num, start_year)
-    response = client.get(dynamic_egdp)
-    assert response.status_code == 200
-
-    # Do the macro job submission
-    response = client.post(dynamic_egdp, egdp_reform)
-    assert response.status_code == 302
-    print(response)
-
-    # The results page will now succeed
-    next_response = client.get(response.url)
-    assert next_response.status_code == 200
-    assert response.url[:-2].endswith("macro_results/")
+    assert "results/" in response.url
     return response