+++ webapp/apps/taxbrain/models.py
@@ -17,7 +17,7 @@ import taxcalc
 import helpers
 import param_formatters
 
-from behaviors import Resultable, Fieldable
+from behaviors import Resultable, Fieldable, DataSourceable
 
 
 # digit or true/false (case insensitive)
@@ -86,7 +86,7 @@ class ErrorMessageTaxCalculator(models.Model):
     text = models.CharField(blank=True, null=False, max_length=4000)
 
 
-class TaxSaveInputs(Fieldable, Resultable, models.Model):
+class TaxSaveInputs(DataSourceable, Fieldable, Resultable, models.Model):
     """
     This model contains all the parameters for the tax model and the tax
     result.